app/requirements.txt
discord.py>=2.5
Pillow>=10.0
aiohttp>=3.9
